Consolidated exception modules
