+++ azure-iot-device/azure/iot/device/__init__.py
@@ -5,9 +5,14 @@ from an IoT device.
 """
 
 from .iothub_session import IoTHubSession  # noqa: F401
-from .iot_exceptions import IoTHubError  # noqa: F401
 from .provisioning_session import ProvisioningSession  # noqa: F401
-from .provisioning_exceptions import ProvisioningServiceError  # noqa: F401
+from .iot_exceptions import (  # noqa: F401
+    IoTHubError,
+    IoTEdgeError,
+    IoTEdgeEnvironmentError,
+    ProvisioningServiceError,
+    IoTHubClientError,
+)
 
 # TODO: Consider not exposing these
 from .mqtt_client import MQTTError, MQTTConnectionFailedError  # noqa: F401
