--- azure-iot-device/azure/iot/device/provisioning_exceptions.py
@@ -1,12 +1,0 @@
-# -------------------------------------------------------------------------
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License. See License.txt in the project root for
-# license information.
-# --------------------------------------------------------------------------
-"""Define Azure IoT domain exceptions to be shared across modules"""
-
-
-class ProvisioningServiceError(Exception):
-    """Represents a failure reported by Provisioning Service"""
-
-    pass
